Build import part of the provider interface
